Moved Terminal to Program.cs
