+++ Echo.Experimental/Program.cs
@@ -5,7 +5,6 @@ using CodeHelpers.Diagnostics;
 using CodeHelpers.Packed;
 using Echo.Common.Mathematics.Randomization;
 using Echo.Core.Aggregation.Acceleration;
-using Echo.Core.Textures.Colors;
 using Echo.Experimental.Benchmarks;
 using JitBuddy;
 
@@ -25,7 +24,8 @@ public class Program
 		// BenchmarkRunner.Run<Loops>();
 		// BenchmarkRunner.Run<AabbSimd>();
 		// BenchmarkRunner.Run<MathFunctions>();
-		BenchmarkRunner.Run<BufferCopy>();
+		// BenchmarkRunner.Run<BufferCopy>();
+		BenchmarkRunner.Run<Timing>();
 	}
 
 	static void TestMonteCarlo()
